sop-skill/python/requirements.txt
pydantic>=2.0,<3.0
jinja2>=3.1,<4.0
chardet>=5.0,<6.0
charset-normalizer>=3.0,<4.0
python-docx>=0.8,<2.0
pymupdf>=1.23,<2.0
openpyxl>=3.1,<4.0
pyyaml>=6.0,<7.0
scikit-learn>=1.0,<2.0
